tests/filesystem: Add capability to run tests on virtiofs filesystem These changes should allow to run tests on virtiofs filesystem. Generally this filesystem is exported by an entitiy outside VM and inside VM we see only a virtiofs device. So setting this up from tests does not seem to be easily possible. Hence I think users will need to setup virtiofs and then run tests manually.
$ cd tests/filesystem/
$ ./test -f virtiofs -t myfs

Here "-f" specifis type of filesystem while "-t" specifies tag to mount.
This will vary depending on under what tag name user has exported a
filesystem to VM.

I have disabled "defcontext" related tests. These test first mount an
fs using "context=" option and create a file and expect newly created
file to remain unlabeled. And they mount fs again and this time with
"defcontext=" and that unlabeled file shows same label as specified
in "defcontext=".

But this does not work with virtiofs. When file is created with "context=",
it still gets labeled and gets the label as specified in "context=". That
means even with "defcontext=" file continues to show the label as on disk
and not label specified in "defcontext=" and test fails. I suspect that
NFS will have similar issues.

So for now, I have disabled "defcontext" related tests.

These tests don't run as part of "make -C tests test" command. But these
will be in tree and user can setup virtiofs manually and run tests manually.

Signed-off-by: Tingting Mao <timao@redhat.com>